Xungchan

+++ src/services/Order.service.ts
@@ -1,10 +1,11 @@
+import ApiError from "../utils/ApiError.js";
 import { orderDao } from "../daos/Order.dao.js";
-import { orderItemDao } from '../daos/OrderItem.dao.js';
 import { variantDao } from "../daos/Variant.dao.js";
 import { CreateOrderDto } from "../dto/Create.dto.js";
+import { orderItemDao } from '../daos/OrderItem.dao.js';
 import { CreateOrderRequest } from '../dto/Request.dto.js';
-import ApiError from "../utils/ApiError.js";
 import { OrderStatus } from "../utils/OrderStatus.enum.js";
+import { OrderQuery, Pagination } from "../utils/Pagination.js";
 import { statusHistoryDao } from './../daos/StatusHistory.dao.js';
 
 const orderService = {
@@ -57,6 +58,31 @@ const orderService = {
         return orderDao.list();
     },
 
+    getOrdersByQuery: async (query: OrderQuery) => {
+
+        const filter: {
+            userId?: string;
+            currentStatus?: string;
+        } = {};
+
+        if (query.userId) filter.userId = query.userId;
+        if (query.status) filter.currentStatus = query.status;
+
+        const page = parseInt((query.page || "1"), 10);
+        const limit = parseInt((query.limit || "10"), 10);
+        const sortBy = query.sortBy || "updatedAt";
+        const sortOrder = query.sortOrder === "desc" ? -1 : 1;
+
+        const options = {
+            skip: (page - 1) * limit,
+            limit: limit,
+            sort: { [sortBy]: sortOrder },
+        }
+        const data = await orderDao.findBy(filter, options);
+        const totalDatas = await orderDao.count(filter);
+        return new Pagination(data, page, limit, totalDatas);
+    },
+
     getOrderById: async (id: string) => {
         const order = await orderDao.readById(id);
         if (!order) {
@@ -71,6 +97,10 @@ const orderService = {
 
     //DELETE
     deleteOrderById: async (id: string) => {
+        const order = await orderDao.readById(id);
+        if (!order) {
+            throw new ApiError(404, "Not Found", "Order not found");
+        }
         return orderDao.deleteById(id);
     },
 
@@ -93,6 +123,11 @@ const orderService = {
         });
         const updated = await orderDao.patchById(id, { currentStatus: status });
         return updated;
+    },
+
+    getRevenue: async () => {
+        const totalRevenue = await orderDao.sumRevenue();
+        return totalRevenue;
     }
 
 };
